+++ packages/kokoas-client/src/pages/projContractSearch/helpers/getCurrentContractStep.ts
