+++ packages/kokoas-client/src/pages/projContractSearch/helpers/latestInvoiceReducer.ts
